setup multiple frontend nodes
