+++ src/tests/simulation/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[target.'cfg(madsim)'.dependencies]
+[dependencies]
 async-trait = "0.1"
 clap = "3"
 glob = "0.3"
